Logic and UI separated in different files
