+++ index.html
@@ -352,7 +352,7 @@
 	</script>
 
 	<!-- Core theme JS-->
-	<script src="./js/shop.js" type="module"></script>
+	<script src="./js/index.js" type="module"></script>
 </body>
 
 </html>
