+++ js/index.js
@@ -1,0 +1,3 @@
+import { printCart } from "./ui";
+
+printCart();
