--- js/shop.js
@@ -1,145 +1,0 @@
-import { products } from "./data.js";
-
-// Improved version of cartList. Cart is an array of products (objects), but each one has a quantity field to define its quantity, so these products are not repeated.
-const cart = [];
-let total = 0;
-
-const buttons = [...document.getElementsByClassName('add-to-cart')]; 
-const cleanCartButton = document.getElementById('clean-cart');
-const countProduct = document.getElementById('count_product');
-
-
-// Exercise 1
-const buy = (id) => {
-    let productToBuy = products.find((product) => product.id === id);
-    let productOnCart = cart.find((product)=> product.id === productToBuy.id);   
-    
-    productOnCart!=undefined?
-        (productOnCart.quantity++)
-        :(cart.push({...productToBuy, quantity: 1}));
-    countProduct.textContent = cart.length;
-    calculateTotal();
-}
-const handleProductAddition = (event) => {
-    let productId = Number(event.target.getAttribute('data-product-id'));
-    buy(productId);
-}
-buttons.forEach((button)=>{
-    button.addEventListener("click", handleProductAddition);
-});
-
-// Exercise 2
-const cleanCart = () =>  {
-    cart.length = 0;
-    total = 0;
-    countProduct.textContent = 0;
-    printCart();
-}
-cleanCartButton.addEventListener('click', cleanCart);
-
-// Exercise 3
-const calculateTotal = () =>  {
-    applyPromotionsCart();
-    total = 0;
-    cart.forEach((product)=>{
-        let productSubtotal = product.subtotalWithDiscount || product.price*product.quantity;
-        total += productSubtotal;
-    });
-}
-
-// Exercise 4
-const applyPromotionsCart = () =>  {
-    cart.forEach(product => {
-        if(product.offer) discountProductPrice(product);
-    });
-}
-
-const discountProductPrice = (product) => {
-    let originalProduct = products.find((originalProduct)=>originalProduct.id === product.id);
-
-    if (product.quantity >= product.offer.number){
-        product.subtotalWithDiscount = (product.price - product.price/100*product.offer.percent) * product.quantity;
-    }else{
-        product.subtotalWithDiscount = null;
-    }
-}
-
-// Exercise 5
-const cartButton = document.querySelector('.cart-button');
-const totalPrice = document.getElementById('total_price');
-const cartList = document.getElementById('cart_list');
-
-const printCart = () => {
-    cartList.innerHTML = '';
-    cart.forEach((product) => {
-        let productRow = document.createElement('tr');
-        productRow.innerHTML = `
-            <tr>
-                <th scope="row">${capitalizeFirstLetter(product.name)}</th>
-                <td>$${(product.price)}</td>
-                <td class='w-2'>
-                    <button type="button" class="btn btn-sm btn-secondary decrement-quantity" product-id='${product.id}'>-</button>
-                    ${product.quantity}
-                    <button type="button" class="btn btn-sm btn-secondary increment-quantity" product-id='${product.id}'>+</button>
-                </td>
-                <td>$${(product.subtotalWithDiscount)?.toFixed(2) || (product.price * product.quantity).toFixed(2)}</td>
-            </tr>
-        `;        
-        cartList.appendChild(productRow);
-    });
-    dynamicDecrementEventListener();    
-    dynamicIncrementEventListener();
-    totalPrice.innerHTML = total.toFixed(2);
-}
-cartButton.addEventListener('click', printCart);
-
-const capitalizeFirstLetter = (val)=>  {
-    return String(val).charAt(0).toUpperCase() + String(val).slice(1);
-}
-
-// ** Nivell II **
-
-// Exercise 7
-let decrementButtons = [];
-
-const removeFromCart = (event) => {
-    let id = Number(event.currentTarget.getAttribute('product-id'));   
-    let product = cart.find((product)=> product.id === id);
-    if(product.quantity === 1){
-        let index = cart.indexOf(product);
-        cart.splice(index,1);
-    }else{
-        product.quantity --;
-    }
-    calculateTotal();
-    printCart();
-}
-const dynamicDecrementEventListener = () => {
-    decrementButtons = [...document.getElementsByClassName('decrement-quantity')];
-    
-    decrementButtons.forEach((button)=>{
-        button.addEventListener('click', removeFromCart);
-    })
-}
-
-// Extra: Increment Button
-let incrementButtons = [];
-
-const addFromCart = (event) => {
-    let id = Number(event.currentTarget.getAttribute('product-id'));   
-    let product = cart.find((product)=> product.id === id);
-    product.quantity ++;
-    calculateTotal();
-    printCart();
-}
-const dynamicIncrementEventListener = () => {
-    incrementButtons = [...document.getElementsByClassName('increment-quantity')];
-    
-    incrementButtons.forEach((button)=>{
-        button.addEventListener('click', addFromCart);
-    })
-}
-
-const open_modal = () =>  {
-    printCart();
-}
