+++ package.json
@@ -1,0 +1,20 @@
+{
+  "name": "s2.2-e-commerce",
+  "version": "1.0.0",
+  "description": "ShopNow is a functional e-commerce web application built with vanilla JavaScript. This project demonstrates modern web development practices including DOM manipulation, shopping cart management, form validation, and discount promotions. Users can browse products across different categories (grocery, beauty, and clothes), add items to their cart, and complete checkout with form validation.",
+  "homepage": "https://github.com/mpujazon/s2.2-e-commerce#readme",
+  "bugs": {
+    "url": "https://github.com/mpujazon/s2.2-e-commerce/issues"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/mpujazon/s2.2-e-commerce.git"
+  },
+  "license": "ISC",
+  "author": "",
+  "type": "commonjs",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  }
+}
